Refactor: move networking monitoring to separate package
