+++ types.go
@@ -17,10 +17,6 @@ func floatToIntPercentRoundUP(f float64) int {
 	return int(f*100 + 0.5)
 }
 
-func floatToIntRoundUP(f float64) int {
-	return int(f + 0.5)
-}
-
 type TimeoutError struct {
 	Origin  string
 	Timeout time.Duration
